requirements.txt
fastapi~=0.115.12
uvicorn~=0.23.2
pandas~=2.1.3
numpy~=1.26.2
scikit-learn~=1.3.2
jinja2
matplotlib
pytest
scikit-surprise>=1.1.1
pydantic>=1.10.0
python-dotenv
google-generativeai
